+++ tests/commands/test_combine_release_packages.py
@@ -1,7 +1,13 @@
 import json
+import sys
+from io import BytesIO, StringIO, TextIOWrapper
+from unittest.mock import patch
+
+import pytest
 
 from ocdskit.cli.__main__ import main
-from tests import assert_streaming, run_streaming
+from ocdskit.exceptions import MissingReleasesWarning
+from tests import assert_streaming, read, run_streaming
 
 
 def test_command(monkeypatch):
@@ -33,3 +39,16 @@ def test_command_publisher(monkeypatch):
 
     package = json.loads(actual)
     assert package['publisher']['name'] == 'Acme Inc.'
+
+
+@patch('sys.stderr', new_callable=StringIO)
+@patch('sys.stdout', new_callable=StringIO)
+def test_command_missing_field(stdout, stderr, monkeypatch):
+    stdin = read('record-package_minimal.json', 'rb')
+
+    with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))):
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'combine-release-packages'])
+        main()
+
+    assert stdout.getvalue() == '{"uri":"","publisher":{"name":"Acme"},"publishedDate":"","version":"1.1","releases":[]}\n'
+    assert stderr.getvalue() == 'item 0 has no "releases" field (check that it is a release package)\n'